# broker/swagger.yaml
swagger: "2.0"
info:
  description: "Vibranium broker"
  version: "1.0.0"
  title: "Swagger Broker"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "robsonkades@outlook.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
tags:
- name: "broker"
  description: "Executa operações no broker"
  externalDocs:
    description: "Find out more"
    url: "http://swagger.io"
schemes:
- "http"
paths:
  /orders:
    post:
      tags:
      - "broker"
      summary: "Cria uma order"
      description: ""
      operationId: "createOrder"
      consumes:
      - "application/json"
      produces:
      - "text/plain"
      parameters:
      - in: "body"
        name: "body"
        description: "Criar uma order de compra ou venda no broker"
        required: true
        schema:
          $ref: "#/definitions/CreateOrder"
      responses:
        "400":
          description: "Invalid input"
        "200":
          description: "Order criada"
  /wallets:
    post:
      tags:
      - "wallet"
      summary: "Cria uma carteria"
      description: ""
      operationId: "createWallet"
      consumes:
      - "application/json"
      produces:
      - "text/plain"
      parameters:
      - in: "body"
        name: "body"
        description: "Criar uma carteira"
        required: true
        schema:
          $ref: "#/definitions/CreateWallet"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Invalid Request"
  /wallets/{wallet}/{symbol}:
    get:
      tags:
      - "wallet"
      summary: "Retorna os valores da carteria"
      description: "Retorna o valores da carteria referente a moeda"
      operationId: "getWalletsByWalletAndSymbol"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "wallet"
        in: "path"
        description: "Carteria"
        required: true
        type: "string"
      - name: "symbol"
        in: "path"
        description: "Moeda/Simbolo"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/WalletResponse"
        "400":
          description: "Invalid Request"
definitions:
  CreateOrder:
    type: "object"
    required:
    - "amount"
    - "price"
    - "wallet"
    - "symbol"
    - "type"
    properties:
      amount:
        type: "number"
        example: 100
      price:
        type: "number"
        example: 100
      wallet:
        type: "string"
        example: "robson-kades-wallet"
      symbol:
        type: "string"
        description: "Moeda/Simbolo"
        enum:
        - "VBN"
      type:
        type: "string"
        description: "Tipo de operação"
        enum:
        - "SELL"
        - "BUY"
  WalletResponse:
    type: "object"
    required:
    - "amount"
    - "price"
    - "wallet"
    - "symbol"
    properties:
      amount:
        type: "number"
        example: 100
      price:
        type: "number"
        example: 100
      wallet:
        type: "string"
        example: "robson-kades-wallet"
      symbol:
        type: "string"
        description: "Vibranium"
        enum:
        - "VBN"
  CreateWallet:
    type: "object"
    required:
    - "amount"
    - "price"
    - "wallet"
    properties:
      amount:
        type: "number"
        example: 100
      price:
        type: "number"
        example: 100
      wallet:
        type: "string"
        example: "robson-kades-wallet"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"